CLASS {xts}	R Documentation

Extract and Set .CLASS Attribute

Description

Extraction and replacement functions to access the xts '.CLASS' attribute. The '.CLASS' attribute is used by reclass() to transform an xts object back to its original class.

Usage

CLASS(x)

CLASS(x) <- value

Arguments

x
An xts object.

value
The new value to assign to the '.CLASS' attribute.

Details

This is meant for use in conjunction with try.xts() and reclass() and is is not intended for daily use. While it's possible to interactively coerce objects to other classes than originally derived from, it's likely to cause unexpected behavior. It is best to use the usual as.xts() and other classes' as methods.

Value

Called for its side-effect of changing the '.CLASS' attribute.

Author(s)

Jeffrey A. Ryan

See Also

as.xts(), reclass()

[Package xts version 0.14.2 Index]