addSeries {xts}	R Documentation

Add a time series to an existing xts plot

Description

Add a time series to an existing xts plot

Usage

addSeries(
  x,
  main = "",
  on = NA,
  type = "l",
  col = NULL,
  lty = 1,
  lwd = 1,
  pch = 1,
  ...
)

Arguments

x
An xts object to add to the plot.

main
Main title for a new panel if drawn.

on
Panel number to draw on. A new panel will be drawn if on = NA.

type
The type of plot to be drawn, same as in plot().

col
Color palette to use, set by default to rational choices.

lty
Set the line type, same as in par().

lwd
Set the line width, same as in par().

pch
The type of plot to be drawn, same as in par().

...
Any other passthrough graphical parameters.

Author(s)

Ross Bennett

[Package xts version 0.14.2 Index]