axTicksByTime {xts}	R Documentation

Compute x-Axis Tickmark Locations by Time

Description

Compute x-axis tickmarks like axTicks in base but with respect to time. Additionally the first argument is the object indexed by time which you are looking to derive tickmark locations for.

It is possible to specify the detail you are seeking, or by passing 'auto' to the ticks.on argument, to get a best heuristic fit.

Usage

axTicksByTime(x, ticks.on='auto', k = 1,
              labels=TRUE, format.labels=TRUE, ends=TRUE,
              gt = 2, lt = 30)

Arguments

x
the object indexed by time, or a vector of times/dates

ticks.on
what to break on

k
frequency of breaks

labels
should a labeled vector be returned

format.labels
format labels - may be format to use

ends
should the ends be adjusted

gt
lower bound on number of breaks

lt
upper bound on number of breaks

Details

This function is written for internal use, and documented for those wishing to use outside of the internal function uses. In general it is most unlikely that the end user will call this function directly.

The format.labels allows for standard formatting like that used in format, strptime, and strftime.

Value

A vector of index points to break on, possibly with the index names.

Author(s)

Jeffrey A. Ryan

See Also

endpoints

Examples

data(sample_matrix)
axTicksByTime(as.xts(sample_matrix),'auto')
axTicksByTime(as.xts(sample_matrix),'weeks')
axTicksByTime(as.xts(sample_matrix),'months',7)

[Package xts version 0.11-0 Index]