na.locf.xts {xts}	R Documentation

Last Observation Carried Forward

Description

xts method replace NA with most recent non-NA

Usage

## S3 method for class 'xts'
na.locf(object, na.rm = FALSE, fromLast = FALSE, maxgap = Inf, ...)

Arguments

object
An xts object.

na.rm
Logical indicating whether leading/trailing NA should be removed. The default is FALSE unlike the zoo method.

fromLast
Logical indicating whether observations should be carried backward rather than forward. Default is FALSE.

maxgap
Consecutive runs of observations more than 'maxgap' will remain NA. See na.locf() for details.

...
Unused.

Details

This is the xts method for the S3 generic na.locf(). The primary difference to note is that after the NA fill action is carried out, the default it to leave trailing or leading NA's in place. This is different than zoo behavior.

Value

An object where each NA in object is replaced by the most recent non-NA prior to it. See na.locf() for details.

Author(s)

Jeffrey A. Ryan

See Also

na.locf()

Examples

x <- xts(1:10, Sys.Date()+1:10)
x[c(1,2,5,9,10)] <- NA

x
na.locf(x)
na.locf(x, fromLast=TRUE)
na.locf(x, na.rm=TRUE, fromLast=TRUE)

[Package xts version 0.14.2 Index]