nseconds {xts}	R Documentation

Number of Periods in Data

Description

Calculate the number of specified periods in a given time series like data object.

Usage

nseconds(x)

nminutes(x)

nhours(x)

ndays(x)

nweeks(x)

nmonths(x)

nquarters(x)

nyears(x)

Arguments

x
A time-based object.

Details

Essentially a wrapper to endpoints() with the appropriate period specified. The result is the number of endpoints found.

As a compromise between simplicity and accuracy, the results will always round up to the nearest complete period. Subtract 1 from the result to get the completed periods.

For finer grain detail one should call the higher frequency functions.

An alternative summary can be found with periodicity(x) and unclass(periodicity(x)).

Value

The number of respective periods in x.

Author(s)

Jeffrey A. Ryan

See Also

endpoints()

Examples

## Not run: 
getSymbols("QQQQ")

ndays(QQQQ)
nweeks(QQQQ)

## End(Not run)

[Package xts version 0.14.2 Index]