period.min {xts}	R Documentation

Calculate Min By Period

Description

Calculate a minimum for each period of INDEX. Essentially a rolling application of minimum over a series of non-overlapping sections.

Usage

period.min(x, INDEX)

Arguments

x
a univariate data object

INDEX
a numeric vector of endpoints to calculate maximum on

Details

Used to calculate a minimum per period given an arbitrary index of sections to be calculated over. This is an optimized function for minimum. There are additional optimized versions for max, sum, and prod.

For xts-coercible objects, an appropriate INDEX can be derived from a call to endpoints.

Value

An xts or zoo object of minimums, indexed by the period endpoints.

Author(s)

Jeffrey A. Ryan

See Also

endpoints, period.sum, period.max, period.prod

Examples

period.min(c(1,1,4,2,2,6,7,8,-1,20),c(0,3,5,8,10))

data(sample_matrix)
period.min(sample_matrix[,1],endpoints(sample_matrix))
period.min(as.xts(sample_matrix)[,1],endpoints(sample_matrix))

[Package xts version 0.10-1 Index]