CLASS {xts}	R Documentation

Extract and Set .CLASS Attribute

Description

Simple extraction and replacement function to access xts .CLASS attribute. The .CLASS attribute is used by reclass to transform an xts object back to its original class.

Usage

CLASS(x)

CLASS(x) <- value

Arguments

x
an xts object

value
the new .CLASS value to assign

Details

It is not recommended that CLASS be called in daily use. While it may be possible to coerce objects to other classes than originally derived from, there is little, if any, chance that the reclass function will perform as expected.

It is best to use the traditional as methods.

Value

Called for its side-effect of changing the .CLASS attribute

Author(s)

Jeffrey A. Ryan

See Also

as.xts,reclass

[Package xts version 0.10-1 Index]