endpoints {xts}	R Documentation

Locate Endpoints by Time

Description

Extract index values of a given xts object corresponding to the last observations given a period specified by on

Usage

endpoints(x, on="months", k=1)

Arguments

x
an xts object

on
the periods endpoints to find as a character string

k
along every k-th element - see notes

Details

endpoints returns a numeric vector corresponding to the last observation in each period specified by on, with a zero added to the beginning of the vector, and the index of the last observation in x at the end.

Valid values for the argument on include: “us” (microseconds), “microseconds”, “ms” (milliseconds), “milliseconds”, “secs” (seconds), “seconds”, “mins” (minutes), “minutes”, “hours”, “days”, “weeks”, “months”, “quarters”, and “years”.

Value

A numeric vector of endpoints beginning with 0 and ending with the a value equal to the length of the x argument.

Author(s)

Jeffrey A. Ryan

Examples

data(sample_matrix)

endpoints(sample_matrix)
endpoints(sample_matrix, 'weeks')

[Package xts version 0.10-1 Index]