firstof {xts}	R Documentation

Create a POSIXct Object

Description

Enable fast creation of time stamps corresponding to the first or last observation in a specified time period.

Usage

firstof(year = 1970, month = 1, day = 1, hour = 0, min = 0, sec = 0, tz = "")

Arguments

year,month,day
numerical values to specify a day

hour,min,sec
numerical vaues to specify time within a day

tz
timezone used for conversion

Details

A wrapper to the R function ISOdatetime with defaults corresponding to the first or last possible time in a given period.

Value

An object of class POSIXct.

Author(s)

Jeffrey A. Ryan

See Also

ISOdatetime

Examples

firstof(2000)
firstof(2005,01,01)

lastof(2007)
lastof(2007,10)

[Package xts version 0.11-0 Index]