periodicity {xts}	R Documentation

Approximate Series Periodicity

Description

Estimate the periodicity of a time-series-like object by calculating the median time between observations in days.

Usage

periodicity(x, ...)

Arguments

x
time-series-like object

...
unused

Details

A simple wrapper to quickly estimate the periodicity of a given data. Returning an object of type periodicity.

This calculates the median number of days between observations as a difftime object, the numerical difference, the units of measurement, and the derived scale of the data as a string.

The time index currently must be of either Date or POSIX class, or coercible to such.

The only list item of note is the scale. This is an estimate of the periodicity of the data in common terms - e.g. 7 day dialy data is best described as ‘weekly’, and would be returned as such.

Possible scale values are:

‘minute’,‘hourly’, ‘daily’,‘weekly’, ‘monthly’,‘quarterly’, and ‘yearly’.

Value

An object containing a list containing the difftime object, frequency, units, and suitable scale.

Note

This function is only a good estimate for the underlying periodicity. If the series is too short, or has no real periodicity, the return values will obviously be wrong. That said, it is quite robust and used internally within xts.

Author(s)

Jeffrey A. Ryan

See Also

difftime

Examples

zoo.ts <- zoo(rnorm(231),as.Date(13514:13744,origin="1970-01-01"))
periodicity(zoo.ts)

[Package xts version 0.10-1 Index]