rbind.xts {xts}	R Documentation

Concatenate Two or More xts Objects by Row

Description

Concatenate or bind by row two or more xts objects along a time-based index.

Usage

## S3 method for class 'xts'
c(...)

## S3 method for class 'xts'
rbind(..., deparse.level = 1)

Arguments

...
objects to bind

deparse.level
not implemented

Details

Implemented in C, these functions bind xts objects by row, resulting in another xts object

There may be non-unique index values in either the original series, or the resultant series.

Identical indexed series are bound in the order or the arguments passed to rbind. See examples.

All objects must have the same number of columns, as well as be xts objects or coercible to such.

rbind and c are aliases.

For traditional merge operations, see merge.xts and cbind.xts.

Value

An xts object with one row per row for each object concatenated.

Note

This differs from rbind.zoo in that non-unique index values are allowed, in addition to the completely different algorithms used internally.

All operations may not behave as expected on objects with non-unique indices. You have been warned.

rbind is a .Primitive function in R. As such method dispatch occurs at the C-level, and may not be consistent with expectations. See the details section of the base function, and if needed call rbind.xts directly to avoid dispatch ambiguity.

Author(s)

Jeffrey A. Ryan

See Also

merge.xts rbind

Examples

x <- xts(1:10, Sys.Date()+1:10)
str(x)

merge(x,x)
rbind(x,x)
rbind(x[1:5],x[6:10])

c(x,x)

# this also works on non-unique index values
x <- xts(rep(1,5), Sys.Date()+c(1,2,2,2,3))
y <- xts(rep(2,3), Sys.Date()+c(1,2,3))

# overlapping indexes are appended
rbind(x,y)
rbind(y,x)

[Package xts version 0.10-1 Index]