xts-package {xts}	R Documentation

xts: extensible time-series

Description

Extensible time series class and methods, extending and behaving like zoo.

Details

Package: 	 xts

Type: 	 Package

Version: 	 0.10-2

Date: 	 2018-03-13

License: 	 GPL (>= 2)

Easily convert one of R's many time-series (and non-time-series) classes to a true time-based object which inherits all of zoo's methods, while allowing for new time-based tools where appropriate.

Additionally, one may use xts to create new objects which can contain arbitrary attributes named during creation as name=value pairs.

Author(s)

Jeffrey A. Ryan and Joshua M. Ulrich

Maintainer: Joshua M. Ulrich <josh.m.ulrich@gmail.com>

See Also

xts as.xts reclass zoo

[Package xts version 0.11-0 Index]